IndexableSlice
